Merged in feature/render-MEI (pull request #7)
Rendering MEI

* Implement frontend MEI rendering

While pretty messy at the moment, this commit now renders MEI files. I'm
going to make this not just happen on the main page, and I want to
abstract this functionality, perhaps into a static JS file, for ease of
reading.

* Style rendered music to fit well

Had to look up some SVG properties to do this well, hehe.

* Abstract rendering process

I have removed the rendering process from the main template, and
implemented JS methods, that are served statically, that can be called
in the template, to render an MEI from a URL to a given div element.
This should hopefully be more clean-looking and reusable.

I have been writing my Javascript to adhere to Standard JS. However, I
haven't discussed this with the team yet, and as such may be prone to
change, awaiting confirmation.

* Make template accept renderUrl

This is to demonstrate how we can make our templates a little more
generic, which is good.

I suspect this will change later on to simply pass the MEI into the
template, given that we have easy database access on the backend.

* Add helper method in JS

Just to help ensure there is only one toolkit per rendering element.

* Render things better

There's a option that can be passed in to make Credo render a
better-scalable SVG.

* Serve Verovio Toolkit on static

This stops us from requesting an external URL.

* Comment better

* Fix style bug

I accidentally stopped the footer from rising upward when it shouldn't.

* Add sidebar for small devices

* Improve mobile display

Also added a title.

* Implement frontend MEI rendering

While pretty messy at the moment, this commit now renders MEI files. I'm
going to make this not just happen on the main page, and I want to
abstract this functionality, perhaps into a static JS file, for ease of
reading.

* Style rendered music to fit well

Had to look up some SVG properties to do this well, hehe.

* Abstract rendering process

I have removed the rendering process from the main template, and
implemented JS methods, that are served statically, that can be called
in the template, to render an MEI from a URL to a given div element.
This should hopefully be more clean-looking and reusable.

I have been writing my Javascript to adhere to Standard JS. However, I
haven't discussed this with the team yet, and as such may be prone to
change, awaiting confirmation.

* Make template accept renderUrl

This is to demonstrate how we can make our templates a little more
generic, which is good.

I suspect this will change later on to simply pass the MEI into the
template, given that we have easy database access on the backend.

* Add helper method in JS

Just to help ensure there is only one toolkit per rendering element.

* Render things better

There's a option that can be passed in to make Credo render a
better-scalable SVG.

* Serve Verovio Toolkit on static

This stops us from requesting an external URL.

* Comment better

* Fix style bug

I accidentally stopped the footer from rising upward when it shouldn't.

* Add sidebar for small devices

* Improve mobile display

Also added a title.

* Abstract sidebar into separate JS file

It had nothing to do with music rendering; I forgot to remove it from
the rendering JS file.

* Rebase onto develop

* Implement songs and song pages

* Implement rendering of editions from database

* Fix bug in song template

Nice work, Craig. Closing a <div> tag with </ul> is unprecedented
genious.

* Move rendering script inclusion to edition file

That way, pages that don't need render functionality aren't slowed down
by unnecessary requirements.

* Remove unused static file

Now that we can render MEI files from the database, we don't need to
serve an MEI file statically.

* Remove test template HTML

We don't use it anymore, and its code has been implemented elsewhere, in
more appropriate places.

* Move rendering script include after footer

This way, the footer shows while the rendering scripts are loading,
instead of it appearing after the scripts are loaded.

* Merge branch 'develop' into feature/render-MEI

* Merged in feature/file-upload (pull request #13)

Switch to FileField objects

* Change BinaryField to FileField on MEI table

This should allow for easier file manipulation, and Django file upload.

* Save mei files to own directory

* Modify mei view to work with new FileField objects

Approved-by: Kpaiy

* Merge branch 'develop' into feature/render-MEI

* Implement Revision page and bug fix

Also did some cleaning up after the merge.

* Clean up migrations

* Change songs.html to song_list.html

* Remove print statement

Approved-by: Joel