Fix code so that adding additional records from Django admin site interface does not cause an error
This originally occured because I specified a primary key id per record. If you try to insert a record
via admin site with a manual key, Django returns an error stating the key already exists (due to manual pk insertion).
Strangely enough if you keep attempting to insert via the site, it increments the pk key it attempts to insert in.

- Condensed code
- Adjusted delete conditions to look for the name of a composer, song, etc. instead of an arbitray pk.