Implement comparison view page
There is now a page to view the comparison. This has a non-functional
'Make Revision' button, and only shows the diff, not the sources, unlike
the compare route that takes an 's' list.

This uses the same route as that compare page, but takes edition and
revision IDs instead of MEI IDs.