Fix revision getting
For some reason, if filtering revisions based on a song or shared
editions, if the same edition/song is listed twice on the revision (e.g.
a revision of two editions for a song), the returned QuerySet would
contain the same record twice. This breaks our use of `.get`. To fix
this, I have replaced `.get` with `.filter` calls, and then used
`.distinct('id')` in order to ensure each revision is only returned
once.